dnd_python_game/requirements.txt
openai>=1.0.0
pyyaml>=6.0
rich>=13.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
websockets>=12.0
aiofiles>=23.0
